+++ src/Elasticsearch/Client.php
@@ -97,7 +97,7 @@ class Client
     /**
      * @param $params array Associative array of parameters
      *
-     * @return array
+     * @return bool
      */
     public function ping($params = [])
     {
@@ -512,7 +512,7 @@ class Client
      *
      * @param $params array Associative array of parameters
      *
-     * @return array
+     * @return array | boolean
      */
     public function exists($params)
     {