+++ src/Elasticsearch/Helper/Iterators/SearchHitIterator.php
@@ -14,7 +14,7 @@ use Iterator;
  * @link     http://elasticsearch.org
  * @see      Iterator
  */
-class SearchHitIterator implements Iterator {
+class SearchHitIterator implements Iterator, \Countable {
 
     /**
      * @var SearchResponseIterator
@@ -35,6 +35,11 @@ class SearchHitIterator implements Iterator {
      * @var array|null
      */
     protected $current_hit_data;
+
+    /**
+     * @var int
+     */
+    protected $count;
 
     /**
      * Constructor
@@ -61,6 +66,11 @@ class SearchHitIterator implements Iterator {
         $current_page = $this->search_responses->current();
         if($this->search_responses->valid() && empty($current_page['hits']['hits'])) {
             $this->search_responses->next();
+        }
+
+        $this->count = 0;
+        if (isset($current_page['hits']) && isset($current_page['hits']['total'])) {
+            $this->count = $current_page['hits']['total'];
         }
 
         $this->readPageData();
@@ -135,5 +145,17 @@ class SearchHitIterator implements Iterator {
             $this->current_hit_data = null;
         }
 
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    public function count()
+    {
+        if ($this->count === null) {
+            $this->rewind();
+        }
+
+        return $this->count;
     }
 }