+++ tests/Elasticsearch/Tests/YamlRunnerTest.php
@@ -110,6 +110,9 @@ class YamlRunnerTest extends \PHPUnit_Framework_TestCase
         // TODO ewwww...
         shell_exec('rm -rf /tmp/test_repo_create_1_loc');
         shell_exec('rm -rf /tmp/test_repo_restore_1_loc');
+        shell_exec('rm -rf /tmp/test_cat_repo_1_loc');
+        shell_exec('rm -rf /tmp/test_cat_repo_2_loc');
+        shell_exec('rm -rf /tmp/test_cat_snapshots_1_loc');
 
         $this->waitForYellow();
     }
@@ -755,12 +758,8 @@ EOF;
     {
         //all_path_options
         $skipList = array(
-            'indices.delete_mapping/all_path_options.yaml',
-            'indices.exists_type/10_basic.yaml',
-            'indices.get_mapping/10_basic.yaml',
-            'indices.create/10_basic.yaml',
-            'indices.get_alias/10_basic.yaml',
-            'cat.allocation/10_basic.yaml'      //regex breaks PHP
+            'cat.nodeattrs/10_basic.yaml',
+            'cat.repositories/10_basic.yaml'
         );
 
         foreach ($skipList as $skip) {